+++ packages/workspace/src/index.ts
@@ -3,3 +3,5 @@ export * from "./knowledgeRoleModel";
 export * from "./tagKnowledgeModel";
 export * from "./themeModel";
 export * from "./shortcutModel";
+
+export * from "./libraryModel";